Services refactored

+++ server/src/controllers/list.ts
@@ -1,10 +1,13 @@
 import { Request, Response, NextFunction } from 'express';
 import * as ListService from '../services/list';
 import * as BoardService from '../services/board';
+import { BoardDocument } from '../models/board';
 
 export const createList = async function (req: Request, res: Response) {
   try {
-    const { board } = req.token;
+    const { board: boardToken } = req.token;
+    const board = await BoardService.find(boardToken);
+    if (!board) throw 'Error creating list';
     const list = ListService.create(req.body.title);
     const update = await ListService.addToBoard(board, list);
     return res.send(update);
@@ -15,37 +18,45 @@ export const createList = async function (req: Request, res: Response) {
 
 export const deleteList = async function (req: Request, res: Response) {
   try {
-    const { board } = req.token;
+    const { board: boardToken } = req.token;
+    const board = await BoardService.find(boardToken);
+    if (!board) throw 'Error deleting list';
     const update = await ListService.remove(board, req.body.listId);
     return res.send(update);
   } catch (err) {
+    console.log(err);
     return res.send(err);
   }
 };
 
 export const updateList = async function (req: Request, res: Response) {
   try {
-    const { board } = req.token;
-    const updateTitle = await ListService.updateTitle(
-      board,
-      req.body.listId,
-      req.body.title
+    const { board: boardToken } = req.token;
+    const board = await BoardService.find(boardToken);
+    const list = await ListService.find(
+      board as BoardDocument,
+      req.body.listId
     );
-    const updateTasks = await ListService.updateTasks(
+    if (!board || !list) throw 'Error updating list';
+
+    const update = await ListService.update(
       board,
-      req.body.listId,
+      list,
+      req.body.title,
       req.body.tasks
     );
-    return res.send({ updateTitle, updateTasks });
+
+    return res.send({ update });
   } catch (err) {
-    console.log(err);
     res.status(404).send(err);
   }
 };
 
 export const updateListOrder = async function (req: Request, res: Response) {
   try {
-    const { board } = req.token;
+    const { board: boardToken } = req.token;
+    const board = await BoardService.find(boardToken);
+    if (!board) throw 'Error updating list order';
     const update = await ListService.reorder(board, req.body.columnOrder);
     return res.send(update);
   } catch (err) {
